+++ build_runner/lib/build_runner.dart
@@ -11,5 +11,14 @@ export 'src/generate/performance_tracker.dart'
     show BuildPerformance, BuildActionPerformance;
 export 'src/generate/phase.dart';
 export 'src/package_builder/package_builder.dart' show PackageBuilder;
+export 'src/package_graph/apply_builders.dart'
+    show
+        BuilderApplication,
+        createBuildActions,
+        apply,
+        toPackage,
+        toAllPackages,
+        toDependentsOf,
+        toPackages;
 export 'src/package_graph/package_graph.dart';
 export 'src/server/server.dart' show ServeHandler;