+++ build_runner/lib/src/package_graph/package_graph.dart
@@ -8,6 +8,8 @@ import 'package:cli_util/cli_util.dart';
 import 'package:path/path.dart' as p;
 import 'package:yaml/yaml.dart';
 
+import 'dependency_ordering.dart';
+
 /// The SDK package, we filter this to the core libs and dev compiler
 /// resources.
 final PackageNode _sdkPackageNode = new PackageNode(
@@ -133,6 +135,7 @@ class PackageGraph {
   ///
   /// See [orderedPackages] for ordering guarantees. The node for [packageName]
   /// will not be included in the result.
+  @Deprecated('Use `createBuildActions` rather than adding actions manually')
   Iterable<PackageNode> dependentsOf(String packageName) {
     if (!allPackages.containsKey(packageName)) return const [];
     var node = allPackages[packageName];
@@ -147,8 +150,11 @@ class PackageGraph {
   /// last. For any two packages for which neither is a transitive dependency of
   /// the other the relative position of the packages within the cycle is
   /// non-deterministic.
+  @Deprecated('Use `createBuildActions` rather than adding actions manually')
   Iterable<PackageNode> get orderedPackages =>
-      _orderedPackages(root, new Set<PackageNode>());
+      stronglyConnectedComponents<String, PackageNode>(
+              [root], (node) => node.name, (node) => node.dependencies)
+          .expand((c) => c);
 
   @override
   String toString() {
@@ -158,16 +164,6 @@ class PackageGraph {
     }
     return buffer.toString();
   }
-}
-
-Iterable<PackageNode> _orderedPackages(
-    PackageNode current, Set<PackageNode> seen) sync* {
-  seen.add(current);
-  for (var dep in current.dependencies) {
-    if (seen.contains(dep)) continue;
-    yield* _orderedPackages(dep, seen);
-  }
-  yield current;
 }
 
 /// A node in a [PackageGraph].